Add: postRecipientMessage function in dummyPost.js
